requirements.txt
# Python dependencies for GitOps deployment scripts
requests>=2.28.0
PyYAML>=6.0
pytest>=7.0.0
responses>=0.23.0
